# Make binding cache accesses allocation-free