--- a/queries/reflect.go
+++ b/queries/reflect.go
@@ -17,17 +17,6 @@ import (
 	"github.com/volatiletech/strmangle"
 )
 
-type colBindingKey struct {
-	typ     reflect.Type
-	colsKey string
-}
-
-var (
-	mut         sync.RWMutex
-	bindingMaps = make(map[colBindingKey][]uint64)
-	structMaps  = make(map[reflect.Type]map[string]uint64)
-)
-
 // Identifies what kind of object we're binding to
 type bindKind int
 
@@ -235,34 +224,9 @@ func bind(rows *sql.Rows, obj interface{}, structType, sliceType reflect.Type, b
 		ptrSlice = reflect.Indirect(reflect.ValueOf(obj))
 	}
 
-	var strMapping map[string]uint64
-	var sok bool
-	var mapping []uint64
-	var ok bool
-
-	colsKey := makeColsKey(structType, cols)
-
-	mut.RLock()
-	mapping, ok = bindingMaps[colsKey]
-	if !ok {
-		if strMapping, sok = structMaps[structType]; !sok {
-			strMapping = MakeStructMapping(structType)
-		}
-	}
-	mut.RUnlock()
-
-	if !ok {
-		mapping, err = BindMapping(structType, strMapping, cols)
-		if err != nil {
-			return err
-		}
-
-		mut.Lock()
-		if !sok {
-			structMaps[structType] = strMapping
-		}
-		bindingMaps[colsKey] = mapping
-		mut.Unlock()
+	mapping, err := getMappingCache(structType).mapping(cols)
+	if err != nil {
+		return err
 	}
 
 	var oneStruct reflect.Value
@@ -441,18 +405,69 @@ func getBoilTag(field reflect.StructField) (name string, recurse bool) {
 	return nameFragment, true
 }
 
-func makeColsKey(typ reflect.Type, cols []string) colBindingKey {
+var (
+	mappingCachesMu sync.Mutex
+	mappingCaches   = make(map[reflect.Type]*mappingCache)
+)
+
+func getMappingCache(typ reflect.Type) *mappingCache {
+	mappingCachesMu.Lock()
+	defer mappingCachesMu.Unlock()
+
+	cache := mappingCaches[typ]
+	if cache != nil {
+		return cache
+	}
+
+	cache = newMappingCache(typ)
+	mappingCaches[typ] = cache
+
+	return cache
+}
+
+type mappingCache struct {
+	typ reflect.Type
+
+	mu          sync.Mutex
+	structMap   map[string]uint64
+	colMappings map[string][]uint64
+}
+
+func newMappingCache(typ reflect.Type) *mappingCache {
+	return &mappingCache{
+		typ:         typ,
+		structMap:   MakeStructMapping(typ),
+		colMappings: make(map[string][]uint64),
+	}
+}
+
+func (b *mappingCache) mapping(cols []string) ([]uint64, error) {
 	buf := strmangle.GetBuffer()
+	defer strmangle.PutBuffer(buf)
+
 	for _, s := range cols {
 		buf.WriteString(s)
+		buf.WriteByte(0)
 	}
-	hash := buf.String()
-	strmangle.PutBuffer(buf)
 
-	return colBindingKey{
-		typ:     typ,
-		colsKey: hash,
+	key := buf.Bytes()
+
+	b.mu.Lock()
+	defer b.mu.Unlock()
+
+	mapping := b.colMappings[string(key)]
+	if mapping != nil {
+		return mapping, nil
 	}
+
+	mapping, err := BindMapping(b.typ, b.structMap, cols)
+	if err != nil {
+		return nil, err
+	}
+
+	b.colMappings[string(key)] = mapping
+
+	return mapping, nil
 }
 
 // Equal is different to reflect.DeepEqual in that it's both less efficient