This is a followup to #706, which removed the cache key creation for the `reflect.Type`, but left the column key as is.

This PR optimizes the binding cache further, eliminating all allocations from the fast path by using the column key's byte slice as the map key (as string conversions in map accesses do not allocate).

Comparing `boilbench` shows a minor memory improvement and no significant change to time:

```
name                      old time/op    new time/op    delta
BoilDelete/boil-4            661ns ± 1%     664ns ± 0%  +0.52%  (p=0.033 n=8+9)
BoilInsert/boil-4           2.21µs ± 2%    2.20µs ± 1%    ~     (p=0.448 n=9+9)
BoilRawBind/boil-4          9.18µs ± 0%    9.17µs ± 1%    ~     (p=0.604 n=9+10)
BoilSelectAll/boil-4        9.06µs ± 1%    9.02µs ± 0%  -0.52%  (p=0.001 n=9+9)
BoilSelectSubset/boil-4     9.72µs ± 1%    9.66µs ± 1%  -0.62%  (p=0.001 n=9+10)
BoilSelectComplex/boil-4    12.3µs ± 1%    12.3µs ± 1%    ~     (p=0.812 n=10+10)
BoilUpdate/boil-4           1.84µs ± 2%    1.83µs ± 1%    ~     (p=0.855 n=10+9)

name                      old alloc/op   new alloc/op   delta
BoilDelete/boil-4             168B ± 0%      168B ± 0%    ~     (all equal)
BoilInsert/boil-4             968B ± 0%      968B ± 0%    ~     (all equal)
BoilRawBind/boil-4          4.30kB ± 0%    4.30kB ± 0%    ~     (all equal)
BoilSelectAll/boil-4        2.93kB ± 0%    2.86kB ± 0%  -2.20%  (p=0.000 n=10+10)
BoilSelectSubset/boil-4     3.07kB ± 0%    3.01kB ± 0%  -2.13%  (p=0.000 n=10+10)
BoilSelectComplex/boil-4    3.99kB ± 0%    3.93kB ± 0%  -1.65%  (p=0.000 n=10+10)
BoilUpdate/boil-4             920B ± 0%      920B ± 0%    ~     (all equal)

name                      old allocs/op  new allocs/op  delta
BoilDelete/boil-4             8.00 ± 0%      8.00 ± 0%    ~     (all equal)
BoilInsert/boil-4             21.0 ± 0%      21.0 ± 0%    ~     (all equal)
BoilRawBind/boil-4            47.0 ± 0%      47.0 ± 0%    ~     (all equal)
BoilSelectAll/boil-4          48.0 ± 0%      47.0 ± 0%  -2.08%  (p=0.000 n=10+10)
BoilSelectSubset/boil-4       52.0 ± 0%      51.0 ± 0%  -1.92%  (p=0.000 n=10+10)
BoilSelectComplex/boil-4      78.0 ± 0%      77.0 ± 0%  -1.28%  (p=0.000 n=10+10)
BoilUpdate/boil-4             16.0 ± 0%      16.0 ± 0%    ~     (all equal)
```

However, in my project, I often query models/tables with many columns, so the eliminating the column key allocation offers a more significant memory improvement:

```
name                   old time/op    new time/op    delta
HandleNop-4               381µs ± 1%     378µs ± 1%  -0.61%  (p=0.010 n=8+8)
HandleNopParallel-4       266µs ± 0%     265µs ± 1%  -0.40%  (p=0.015 n=8+8)
HandleCustomCommand-4     391µs ±10%     379µs ± 1%    ~     (p=0.277 n=9+8)
HandleMixed-4             391µs ± 5%     392µs ± 1%    ~     (p=0.780 n=9+10)

name                   old alloc/op   new alloc/op   delta
HandleNop-4              15.0kB ± 0%    14.2kB ± 0%  -5.14%  (p=0.000 n=9+9)
HandleNopParallel-4      15.1kB ± 0%    14.3kB ± 0%  -5.12%  (p=0.000 n=9+10)
HandleCustomCommand-4    15.0kB ± 0%    14.2kB ± 0%  -5.12%  (p=0.000 n=10+9)
HandleMixed-4            15.1kB ± 0%    14.3kB ± 0%  -5.13%  (p=0.000 n=10+7)

name                   old allocs/op  new allocs/op  delta
HandleNop-4                 259 ± 0%       258 ± 0%  -0.39%  (p=0.000 n=10+10)
HandleNopParallel-4         259 ± 0%       258 ± 0%  -0.39%  (p=0.000 n=10+10)
HandleCustomCommand-4       259 ± 0%       258 ± 0%  -0.39%  (p=0.000 n=10+10)
HandleMixed-4               260 ± 0%       259 ± 0%  -0.38%  (p=0.000 n=10+10)
```

That's 5% fewer bytes for a single allocation.